+++ src/algebra/hop.rs
@@ -59,45 +59,44 @@ impl <T> HopProgram<T>
     }
   }
 
-  // pub fn delete(&mut self, link: Link) -> HopEffect<()> {
-  //   let link_pairs = self.get_link_pairs()?;
+  pub fn delete(&self, link: Link) -> HopEffect<()> {
+    let link_pairs = self.get_link_pairs()?;
 
-  //   match link_pairs.iter().find(|lp| lp.link == link) {
-  //    Some(pair) => {
-  //        let prompt_message = format!("Are you sure you want to delete {} which links to {} ?", pair.link, pair.target);
+    match link_pairs.iter().find(|lp| lp.link == link) {
+     Some(pair) => {
+         let prompt_message = format!("Are you sure you want to delete {} which links to {} ?", pair.link, pair.target);
 
-  //        let no_action = || {
-  //         &self.value.println(&format!("Aborting delete of {}", pair.link));
-  //         Ok(())
-  //       };
+         let no_action = || {
+          self.value.println(&format!("Aborting delete of {}", &pair.link));
+          Ok(())
+        };
 
-  //        let yes_action = || {
-  //            let hop_home = &self.value.get_hop_home(&self.cfg_dir)?;
-  //            let file_path = (hop_home.clone()).join(&link);
-  //            fs::remove_file(file_path)?;
-  //            // &self.value.println(&format!("Removed link {} which pointed to {}", &link, &pair.target));
-  //            Ok(())
-  //        };
+         let yes_action = || {
+             let hop_home = &self.value.get_hop_home(&self.cfg_dir)?;
+             let delete_result = self.value.delete_link(&hop_home, &pair)?;
 
-  //        self.prompt_user(&prompt_message, yes_action, no_action)
-  //    },
+             self.value.println(&format!("Removed link {} which pointed to {}", &link, &pair.target));
+             Ok(())
+         };
 
-  //    None => Err(io_error(&format!("Could not find link named:{} to delete", link)))
-  //   }
-  // }
+         self.prompt_user(&prompt_message, yes_action, no_action)
+     },
+
+     None => Ok(self.value.println(&format!("Could not find link named:`{}` for deletion", link)))
+    }
+  }
 
   fn prompt_user<Y, N, R>(&self, message: &str, yes_action: Y, no_action: N) -> HopEffect<R> where
       Y: FnOnce() -> HopEffect<R>,
       N: FnOnce() -> HopEffect<R>
   {
-      // self.value.println(message);
-      // let buffer = self.value.readln()?;
-      // let response = buffer.lines().next().ok_or(io_error("Could not retrieve lines from stdio"))?;
-      // match response {
-      //     "Y" | "y"  => yes_action(),
-      //     _ => no_action()
-      // }
-      todo!()
+      self.value.println(message);
+      let buffer = self.value.readln()?;
+      let response = buffer.lines().next().ok_or(io_error("Could not retrieve lines from stdio"))?;
+      match response {
+          "Y" | "y"  => yes_action(),
+          _ => no_action()
+      }
   }
 
 }
@@ -106,7 +105,7 @@ impl <T> HopProgram<T>
 #[cfg(test)]
 mod tests {
     use crate::algebra::{std_io::StdIO, user_dirs::UserDirs, directories::Directories};
-    use crate::algebra::symlinks::{SymLinks, SymLink, SymLinkDeleteStatus};
+    use crate::algebra::symlinks::{SymLinks, SymLink};
     use crate::models::{HopEffect, LinkPair, Link};
     use super::HopProgram;
 
@@ -171,7 +170,7 @@ mod tests {
         Ok(self.link_exists)
       }
 
-      fn delete_link(&self, dir_path: &PathBuf, linkPair: &LinkPair) -> HopEffect<SymLinkDeleteStatus> {
+      fn delete_link(&self, dir_path: &PathBuf, linkPair: &LinkPair) -> HopEffect<()> {
         todo!()
       }
